Add support for scanned images
Scanned images can be uploaded through the API. We store the image as a
file, and we store metadata in the database. We can retrieve an
individual image or the metadata for all of the images.

The server also serves up static files, so that we can present a simple
file upload page.